# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Master API
  description: |
    The Task Master API provides comprehensive project and task management capabilities with multi-tenant organization support.

    ## Features
    - Multi-tenant architecture with organization management
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Comprehensive task and project management
    - Real-time capabilities with webhooks
    - Advanced filtering and search

    ## Authentication
    All endpoints require authentication using Bearer tokens in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```

    ## Rate Limiting
    - Authentication endpoints: 10 requests per minute per IP
    - General API endpoints: 1000 requests per hour per user
    - Bulk operations: 100 requests per hour per user
  version: 1.0.0
  contact:
    name: Task Master API Support
    email: api-support@taskmaster.com
    url: https://docs.taskmaster.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.taskmaster.com/v1
    description: Production server
  - url: http://localhost:3001/api/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authentication
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100

    # User and authentication schemas
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '850e8400-e29b-41d4-a716-446655440003'
        organizationId:
          type: string
          format: uuid
          example: '650e8400-e29b-41d4-a716-446655440001'
        name:
          type: string
          example: 'Website Redesign'
        description:
          type: string
          nullable: true
          example: 'Complete redesign of the company website'
        priority:
          type: string
          enum: [low, medium, high]
          example: 'high'
        status:
          type: string
          enum: [planning, active, completed, cancelled]
          example: 'active'
        dueDate:
          type: string
          format: date-time
          nullable: true
          example: '2024-06-30T23:59:59Z'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-10T00:00:00Z'
    ProjectWithStats:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            taskCount:
              type: integer
              example: 15
            completedTaskCount:
              type: integer
              example: 5
            memberCount:
              type: integer
              example: 3
            progress:
              type: number
              format: float
              example: 33.33

    # Task schemas
...
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          example: 'Website Redesign'
        description:
          type: string
          maxLength: 1000
          example: 'Complete redesign of the company website'
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
          example: 'high'
        dueDate:
          type: string
          format: date-time
          example: '2024-06-30T23:59:59Z'
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          example: 'Website Redesign - Phase 2'
        description:
          type: string
          maxLength: 1000
          example: 'Updated description'
        priority:
          type: string
          enum: [low, medium, high]
          example: 'medium'
        status:
          type: string
          enum: [planning, active, completed, cancelled]
          example: 'active'
        dueDate:
          type: string
          format: date-time
          example: '2024-07-15T23:59:59Z'
...
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Organization ID
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Project ID
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Search term
paths:
...
  /organizations/{organizationId}/projects:
    get:
      tags:
        - Projects
      summary: List projects
      description: Get all projects in an organization.
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [planning, active, completed, cancelled]
          description: Filter by status
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high]
          description: Filter by priority
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          projects:
                            type: array
                            items:
                              $ref: '#/components/schemas/ProjectWithStats'

    post:
      tags:
        - Projects
      summary: Create project
      description: Create a new project in an organization.
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          project:
                            $ref: '#/components/schemas/Project'
  /projects/{projectId}:
    get:
      tags:
        - Projects
      summary: Get project
      description: Get detailed information about a specific project.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          project:
                            $ref: '#/components/schemas/Project'
                          statistics:
                            type: object
                            properties:
                              taskCount:
                                type: integer
                              completedTaskCount:
                                type: integer
                              memberCount:
                                type: integer
                              progress:
                                type: number
                                format: float
                          members:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: string
                                  format: uuid
                                fullName:
                                  type: string
                                role:
                                  type: string

    put:
      tags:
        - Projects
      summary: Update project
      description: Update project information.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          project:
                            $ref: '#/components/schemas/Project'

    delete:
      tags:
        - Projects
      summary: Delete project
      description: Delete a project (admin only).
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: 'Project has been deleted successfully'

  # Task endpoints
...